XW-1940 | Move some files from RP up after merging repos
