+++ bower.json
@@ -7,6 +7,7 @@
   },
   "ignore": [
     "**/.*",
-    "gulpfile.js"
+    "gulpfile.js",
+    "reference-page"
   ]
 }
