+++ gulpfile.js
@@ -1,11 +1,20 @@
-var cheerio = require('gulp-cheerio'),
+var autoprefixer = require('gulp-autoprefixer'),
+	cheerio = require('gulp-cheerio'),
+	clean = require('gulp-clean'),
 	fs = require('fs'),
 	gulp = require('gulp'),
+	inject = require('gulp-inject'),
+	livereload = require('gulp-livereload'),
 	path = require('path'),
 	rename = require('gulp-rename'),
+	scss = require('gulp-sass'),
+	styledown = require('gulp-styledown'),
 	svgmin = require('gulp-svgmin'),
-	svgstore = require('gulp-svgstore');
+	svgstore = require('gulp-svgstore'),
+	watch = require('gulp-watch');
 
+
+// BUILD SVGs
 function renameSvgFiles(folder) {
 	return rename(function (filePath) {
 		// Use `id="wds-company-logo-wikia"` for company/logo-wikia.svg
@@ -82,4 +91,96 @@ gulp.task('svg-individual', function () {
 
 gulp.task('svg', ['svg-sprite', 'svg-individual']);
 
-gulp.task('default', ['svg']);
+
+// BUILD REFERENCE PAGE
+
+var tmpDir = './reference-page/tmp/',
+	outputDir = './reference-page/',
+	outputFile = 'index.html';
+
+function injectFileContents(src, dest, prefix, suffix) {
+	prefix = prefix || '';
+	suffix = suffix || '';
+
+	return gulp.src(tmpDir + outputFile)
+		.pipe(inject(
+			gulp.src(src),
+			{
+				transform: function (filePath, file) {
+					return prefix + file.contents.toString('utf8') + suffix;
+				}
+			}
+		))
+		.pipe(gulp.dest(dest));
+}
+
+// BUILD
+gulp.task('clean-up', function () {
+	return gulp.src(tmpDir, {read: false}).pipe(clean());
+});
+
+gulp.task('build-styles', ['clean-up'], function () {
+	return gulp.src('reference-page/styles/index.scss')
+		.pipe(scss())
+		.pipe(autoprefixer({
+			browsers: ['last 3 versions'],
+			cascade: false
+		}))
+		.pipe(rename('component-styles.css'))
+		.pipe(gulp.dest(tmpDir));
+});
+
+gulp.task('build-html', ['build-styles'], function () {
+	return gulp.src(['./reference-page/styles/*.scss', './reference-page/components/*.md'])
+		.pipe(styledown({
+			config: './reference-page/config/config.md',
+			filename: outputFile
+		}))
+		.pipe(gulp.dest(tmpDir));
+});
+
+gulp.task('inject-styles', ['build-html'], function () {
+	return injectFileContents(
+		['./reference-page/config/default-styles.css', './reference-page/config/custom-styles.css', tmpDir + 'component-styles.css'],
+		tmpDir, '<style>', '</style>'
+	);
+});
+
+gulp.task('inject-scripts', ['inject-styles'], function () {
+	return injectFileContents('./reference-page/config/build.js', tmpDir, '<script>', '</script>');
+});
+
+gulp.task('inject-icons', ['inject-scripts'], function () {
+	return injectFileContents('./dist/svg/*.svg', tmpDir);
+});
+
+gulp.task('build-reference-page', ['inject-icons'], function () {
+	return gulp.src(tmpDir + outputFile).pipe(gulp.dest(outputDir));
+});
+
+// WATCH
+gulp.task('rebuild', ['default'], function (done) {
+	console.log('\n# Finished rebuilding files.');
+	livereload.reload();
+	done();
+});
+
+gulp.task('watch', ['default'], function () {
+	console.log('\n# Gulp is watching your files.\n# You can start developing.');
+	livereload.listen();
+
+	return watch([
+		'**/*.scss',
+		'./assets/**/*.svg',
+		'./reference-page/components/*.md',
+		'./reference-page/config/*.css',
+		'./reference-page/config/*.js',
+		'./reference-page/config/*.md',
+	], function () {
+		console.log('\n# Change detected, rebuilding files...');
+		gulp.start('rebuild');
+	});
+});
+
+// DEFAULT TASK
+gulp.task('default', ['svg', 'build-reference-page']);
