--- reference-page/.gitignore
@@ -1,17 +1,0 @@
-.idea/
-bower_components/
-node_modules/
-tmp/
-
-# Visual Regression Testing files that shouldn't be commited as they are of no use
-visuals/reference-page/diff
-visuals/reference-page/*.regression.png
-visuals/reference-page/reference-page.png
-visuals/reference-page/reference-page.*px.png
-
-npm-debug.log
-*.DS_Store
-*._.DS_Store
-
-# Jenkins
-debug.log
