--- reference-page/bower.json
@@ -1,10 +1,0 @@
-{
-  "name": "design-system-reference-page",
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/Wikia/design-system-reference-page.git"
-  },
-  "dependencies": {
-    "design-system": "Wikia/design-system#0.4.0"
-  }
-}
