+++ reference-page/index.html
@@ -399,7 +399,6 @@ code.sg:first-child {
   display: inline-block;
   position: relative; }
   .wds-dropdown__toggle {
-    pointer-events: none;
     position: relative;
     z-index: 2; }
   .wds-dropdown__toggle-chevron {
