--- reference-page/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "design-system-reference-page",
-  "version": "1.0.0",
-  "scripts": {
-    "setup": "npm install && bower update",
-    "link": "cd ../design-system && bower link && cd ../design-system-reference-page && bower link design-system",
-    "unlink": "rm -rf bower_components/design-system && bower update",
-    "build": "gulp",
-    "dev": "gulp watch",
-    "test": "node test",
-    "update-baseline": "rm -rf visuals/* && npm run test"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Wikia/design-system-reference-page.git"
-  },
-  "devDependencies": {
-    "chai": "3.5.0",
-    "file-url": "1.1.0",
-    "gulp": "3.9.1",
-    "gulp-autoprefixer": "3.1.0",
-    "gulp-clean": "0.3.2",
-    "gulp-inject": "4.1.0",
-    "gulp-livereload": "3.8.1",
-    "gulp-rename": "1.2.2",
-    "gulp-sass": "2.3.2",
-    "gulp-styledown": "0.0.2",
-    "gulp-watch": "4.3.9",
-    "mocha": "2.5.3",
-    "styledown": "1.0.2",
-    "vrt": "Wikia/visual-regression-testing#0.0.4"
-  }
-}
