+++ reference-page/styles/index.scss
@@ -1,4 +1,4 @@
-@import '../bower_components/design-system/dist/styles';
+@import '../../dist/styles';
 
 @import 'breakpoints';
 @import 'colors';
